More meddling...
Alright. This script works now. It is now capable of taking in a single
plt track from Ozi Explorer and pooping out a single point feature
classshape for UTM zone 12N in NAD 83. Further updates will be meant to
clean this, so that it can be treated as a single class!